pyproject.toml
[tool.poetry]
name = "ai-spider"
version = "0.2.0"
description = "ai-worker coordination hub and inference endpoint provider"
authors = ["erik aronesty <erik@q32.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.103.1"
websockets = "^11.0.3"
python-dotenv = "^1.0.0"
sse-starlette = "^1.6.5"
uvicorn = "^0.23.2"
itsdangerous = "^2.1.2"
httpx = "*"
python-multipart = "^0.0.6"
gunicorn = "^21.2.0"
notanorm = "^3.9.1"
pymysql = "^1.1.0"
aioboto3 = "^11.3.1"
asyncio-atexit = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
ai-worker = {git = "ssh://git@github.com/ArcadeLabsInc/ai-worker.git"}
httpx-sse = "^0.3.1"
moto = {extras = ["s3"], version = "^4.2.4"}
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
ignore = ['E701', 'E501']

[tool.ruff.per-file-ignores]
"__init__.py" = ['F401']
"tests/*" = ['E402']
